+++ libyul/AsmAnalysis.h
@@ -59,7 +59,6 @@ public:
 	explicit AsmAnalyzer(
 		AsmAnalysisInfo& _analysisInfo,
 		langutil::ErrorReporter& _errorReporter,
-		boost::optional<langutil::Error::Type> _errorTypeForLoose,
 		Dialect const& _dialect,
 		ExternalIdentifierAccess::Resolver const& _resolver = ExternalIdentifierAccess::Resolver(),
 		std::set<YulString> const& _dataNames = {}
@@ -68,7 +67,6 @@ public:
 		m_info(_analysisInfo),
 		m_errorReporter(_errorReporter),
 		m_dialect(_dialect),
-		m_errorTypeForLoose(_errorTypeForLoose),
 		m_dataNames(_dataNames)
 	{
 		if (EVMDialect const* evmDialect = dynamic_cast<EVMDialect const*>(&m_dialect))
@@ -85,9 +83,7 @@ public:
 	bool operator()(Literal const& _literal);
 	bool operator()(Identifier const&);
 	bool operator()(FunctionalInstruction const& _functionalInstruction);
-	bool operator()(Label const& _label);
 	bool operator()(ExpressionStatement const&);
-	bool operator()(StackAssignment const&);
 	bool operator()(Assignment const& _assignment);
 	bool operator()(VariableDeclaration const& _variableDeclaration);
 	bool operator()(FunctionDefinition const& _functionDefinition);
@@ -110,12 +106,8 @@ private:
 
 	Scope& scope(Block const* _block);
 	void expectValidType(std::string const& type, langutil::SourceLocation const& _location);
-	void warnOnInstructions(dev::eth::Instruction _instr, langutil::SourceLocation const& _location);
-
-	/// Depending on @a m_flavour and @a m_errorTypeForLoose, throws an internal compiler
-	/// exception (if the flavour is not Loose), reports an error/warning
-	/// (if m_errorTypeForLoose is set) or does nothing.
-	void checkLooseFeature(langutil::SourceLocation const& _location, std::string const& _description);
+	bool warnOnInstructions(dev::eth::Instruction _instr, langutil::SourceLocation const& _location);
+	bool warnOnInstructions(std::string const& _instrIdentifier, langutil::SourceLocation const& _location);
 
 	int m_stackHeight = 0;
 	yul::ExternalIdentifierAccess::Resolver m_resolver;
@@ -127,7 +119,6 @@ private:
 	langutil::ErrorReporter& m_errorReporter;
 	langutil::EVMVersion m_evmVersion;
 	Dialect const& m_dialect;
-	boost::optional<langutil::Error::Type> m_errorTypeForLoose;
 	/// Names of data objects to be referenced by builtin functions with literal arguments.
 	std::set<YulString> m_dataNames;
 	ForLoop const* m_currentForLoop = nullptr;