+++ libyul/backends/wasm/EWasmCodeTransform.cpp
@@ -108,21 +108,9 @@ wasm::Expression EWasmCodeTransform::operator()(Assignment const& _assignment)
 	return generateMultiAssignment(move(variableNames), visit(*_assignment.value));
 }
 
-wasm::Expression EWasmCodeTransform::operator()(StackAssignment const&)
-{
-	yulAssert(false, "");
-	return {};
-}
-
 wasm::Expression EWasmCodeTransform::operator()(ExpressionStatement const& _statement)
 {
 	return visitReturnByValue(_statement.expression);
-}
-
-wasm::Expression EWasmCodeTransform::operator()(Label const&)
-{
-	yulAssert(false, "");
-	return {};
 }
 
 wasm::Expression EWasmCodeTransform::operator()(FunctionalInstruction const& _f)