+++ test/libsolidity/syntaxTests/inlineAssembly/invalid/no_identifier_assignment.sol
@@ -1,9 +1,0 @@
-contract C {
-  function f() public pure {
-    assembly {
-      return := 1
-    }
-  }
-}
-// ----
-// ParserError: (70-72): Variable name must precede ":=" in assignment.